requirements.txt
tensorflow>=2.3
numpy
opencv-python
streamlit
pillow
